+++ src/GSettingsDlg.cpp
@@ -10,6 +10,7 @@
 #include <Gtk/Label.hpp>
 #include <Gtk/ToggleButton.hpp>
 #include <Gtk/Scale.hpp>
+#include <Gtk/SpinButton.hpp>
 #include <Gio/SettingsSchemaKey.hpp>
 
 #ifdef GIR_INLINE
@@ -18,6 +19,8 @@
 #include <Gtk/Button.ipp>
 #include <Gtk/Label.ipp>
 #include <Gtk/Scale.ipp>
+#include <Gtk/SpinButton.ipp>
+#include <Gtk/Adjustment.ipp>
 #include <Gio/SettingsSchema.ipp>
 #include <Gio/SettingsSchemaKey.ipp>
 #include <GLib/Variant.ipp>
@@ -69,6 +72,23 @@ void HandleSmoothScroll(Gtk::Builder builder)
     });
 }
 
+void HandleCellHeightAdjustment(Gtk::Builder builder, GFont &font)
+{
+    Gtk::SpinButton spin{builder.get_object("cell_height_adjustment").g_obj()};
+    auto key = GConfig::GetSettingsSchema().get_key(GConfig::CELL_HEIGHT_ADJUSTMENT_KEY);
+    auto range = key.get_range().get_child_value(1);
+    range = range.get_child_value(0);
+    double low = range.get_child_value(0).get_double();
+    double high = range.get_child_value(1).get_double();
+    spin.get_adjustment().set_lower(low);
+    spin.get_adjustment().set_upper(high);
+    spin.set_value(GConfig::GetCellHeightAdjustment());
+    spin.on_value_changed(spin, [=, &font](Gtk::SpinButton spin) {
+        GConfig::SetCellHeightAdjustment(spin.get_value());
+        font.OnChanged();
+    });
+}
+
 } //namespace;
 
 GSettingsDlg::GSettingsDlg(gir::Gtk::Window window, GFont &font)
@@ -78,6 +98,7 @@ GSettingsDlg::GSettingsDlg(gir::Gtk::Window window, GFont &font)
 
     HandleFont(builder, dlg, font);
     HandleSmoothScroll(builder);
+    HandleCellHeightAdjustment(builder, font);
 
     dlg.set_transient_for(window);
     dlg.on_destroy(dlg, [builder](Gtk::Dialog) {